requirements.txt
box2d-py==2.3.8
gym==0.21.0
imageio-ffmpeg==0.4.5
matplotlib==3.5.1
neat-python==0.92
numpy==1.22.2
pyglet==1.5.21
